+++ igel/__init__.py
@@ -1,7 +1,7 @@
 """Top-level package for igel."""
-from .igel import IgelModel
+from .igel import Igel, models_dict, metrics_dict
 
 
 __author__ = "Nidhal Baccouri"
 __email__ = 'nidhalbacc@gmail.com'
-__version__ = '0.1.1'
+__version__ = '0.2.3'