+++ HISTORY.rst
@@ -2,7 +2,62 @@
 History
 =======
 
-0.0.1 (2020-09-05)
+0.2.3 (2020-09-26)
+-------------------
+
+* added clustering support
+* all clustering classes of sklearn are now supported
+
+0.2.2 (2020-09-23)
+-------------------
+
+* added init command
+* users can get started quickly by getting a default yaml file "on the fly"
+
+0.2.1 (2020-09-21)
+-------------------
+
+* added support for other ensemble models like adaboost, extra trees etc..
+
+
+0.2.0 (2020-09-19)
+-------------------
+
+* added the experiment command
+* fixed bugs in predict
+* fixed sphinx docs dependencies
+
+0.1.9 (2020-09-18)
+-------------------
+
+* added support for multioutput regression and classification
+* fixed bug in preparing predict data
+* provided a way to evaluate models
+* changed class from IgelModel to Igel
+* updated igel cli
+
+0.1.8 (2020-09-13)
+------------------
+* fixed predict function bugs and added examples
+
+0.1.7 (2020-09-12)
+------------------
+* implemented optional arguments in sklearn models
+
+
+0.1.5 (2020-09-10)
+------------------
+* implemented encoding and scaling methods
+
+0.1.4 (2020-09-08)
+------------------
+* support for all sklearn models
+
+0.1.3 (2020-09-07)
+------------------
+* implemented basic dataset operations
+
+0.1.0 (2020-09-05)
 ------------------
 * stable release with an end to end pipeline